Changes to FK to make alembic migration backwards compatible
