+++ src/backend/database_models/agent.py
@@ -29,7 +29,9 @@ class Agent(Base):
         nullable=False,
     )
 
-    user_id: Mapped[str] = mapped_column(ForeignKey("users.id", ondelete="CASCADE"))
+    user_id: Mapped[str] = mapped_column(
+        ForeignKey("users.id", ondelete="CASCADE"), nullable=True
+    )
     organization_id: Mapped[Optional[str]] = mapped_column(
         ForeignKey(
             "organizations.id", name="agents_organization_id_fkey", ondelete="CASCADE"
