+++ src/backend/database_models/agent_tool_metadata.py
@@ -10,7 +10,9 @@ from backend.database_models.base import Base
 class AgentToolMetadata(Base):
     __tablename__ = "agent_tool_metadata"
 
-    user_id: Mapped[str] = mapped_column(ForeignKey("users.id", ondelete="CASCADE"))
+    user_id: Mapped[str] = mapped_column(
+        ForeignKey("users.id", ondelete="CASCADE"), nullable=True
+    )
     agent_id: Mapped[str] = mapped_column(
         ForeignKey("agents.id", ondelete="CASCADE"), nullable=False
     )
